feat(ai): restructure AI integration and update imports
