+++ bridge/package.json
@@ -7,21 +7,28 @@
     "dev": "ts-node-dev --respawn --transpile-only src/index.ts",
     "start": "node dist/index.cjs",
     "build": "tsc --project tsconfig.json && esbuild src/index.ts --bundle --platform=node --outfile=dist/index.cjs --format=cjs --packages=external",
+    "build:bundle-for-pkg": "esbuild src/index.ts --bundle --platform=node --outfile=dist/index.bundled.cjs --format=cjs --external:better-sqlite3 --external:@napi-rs/keyring --external:ssh2 --external:cpu-features --external:pg-native",
     "copy:native": "node scripts/copy-native.js",
     "rebuild:native": "npm rebuild better-sqlite3",
-    "build:pkg:win": "npm run build && npm run rebuild:native && npm run copy:native && npx @yao-pkg/pkg . --target node22-win-x64 --output ../src-tauri/resources/bridge-x86_64-pc-windows-msvc.exe",
-    "build:pkg:linux": "npm run build && npm run rebuild:native && npm run copy:native && npx @yao-pkg/pkg . --target node22-linux-x64 --output ../src-tauri/resources/bridge-x86_64-unknown-linux-gnu",
+    "build:pkg:win": "npm run build && npm run build:bundle-for-pkg && npm run rebuild:native && npm run copy:native && npx @yao-pkg/pkg dist/index.bundled.cjs --target node22-win-x64 --output ../src-tauri/resources/bridge-x86_64-pc-windows-msvc.exe",
+    "build:pkg:linux": "npm run build && npm run build:bundle-for-pkg && npm run rebuild:native && npm run copy:native && npx @yao-pkg/pkg dist/index.bundled.cjs --target node22-linux-x64 --output ../src-tauri/resources/bridge-x86_64-unknown-linux-gnu",
     "test": "jest",
     "test:watch": "jest --watchAll --detectOpenHandles"
   },
   "dependencies": {
+    "@anthropic-ai/sdk": "^0.100.1",
+    "@google/generative-ai": "^0.24.1",
     "@jest/globals": "^30.2.0",
+    "@mistralai/mistralai": "^2.2.5",
     "@napi-rs/keyring": "^1.2.0",
     "@types/ssh2": "^1.15.5",
     "bcryptjs": "^3.0.3",
     "better-sqlite3": "^11.9.0",
     "dotenv": "^17.2.3",
+    "groq-sdk": "^1.2.1",
     "mysql2": "^3.15.3",
+    "ollama": "^0.6.3",
+    "openai": "^6.41.0",
     "pg": "^8.16.3",
     "pg-query-stream": "^4.10.3",
     "pino": "^9.14.0",
